+++ src/Krystal/Text/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Added `TextUtils::normalizeColumn()` to convert column names like `first_name` to `First Name`
  * In `Text\SlugGenerator` added `getUniqueSlug()` method to handle uniqueness
  * In `Text\CollectionManager` added `isEmpty()` method that determines whether collection is empty
  * Added `TextUtils::uniqueString()` to generate a very unique string based on MD5